+++ inst/doc/Educacion.R
@@ -1,12 +1,13 @@
 ## ---- include = FALSE---------------------------------------------------------
+options(rmarkdown.html_vignette.check_title = FALSE)
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   dpi = 400
 )
 
-## ----setup--------------------------------------------------------------------
-library(ech)
+## ----eval = FALSE-------------------------------------------------------------
+#  library(ech)
 
 ## ----eval = FALSE-------------------------------------------------------------
 #  df <- age_groups(df)